requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
